Adjusted some visuals. Refined and added on to the code dictating the logic of the color system.
We can now set a tolerance for color sensitive objects, and there’s an option to give the object a transparent texture if the trigger color isn’t the active color. Additionally, I added the option to negate all the logic so that the trigger color turns into a deactivation color. This can all be changed in the editor which makes the script pretty dynamic.
Added a working 3d model of the color swapper.